kernel - Fix buffer cache deadlock due to multiple buffer cache sizes
* The buffer daemon was triggering off of dirtybufspace[hw], comparing
  it against bufspace.  However, bufspace is an aggregation of the
  total buffer kva reserved, not the dirty count based on b_bufsize.

* Add dirtykvasize to track the dirty space in terms of the kva reservation
  and change the buffer daemons to use that instead.

* systat and other displays still use the bufsize, which is a more
  accurate indicator for human consumption.

+++ sys/kern/kern_iosched.c
@@ -159,8 +159,8 @@ bwillread(int bytes)
 void
 bwillinode(int n)
 {
-	int count;
-	int factor;
+	long count;
+	long factor;
 
 	count = bd_heatup();
 	if (count > 0) {
